refactor: extract steam API and helpers to separate libraries
