+++ apps/perion-steam-challenge/env.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+    interface ProcessEnv {
+        STEAM_API_KEY: string;
+    }
+}
