+++ apps/perion-steam-challenge/tsconfig.json
@@ -29,7 +29,8 @@
     "../../apps/perion-steam-challenge/.next/types/**/*.ts",
     "../../dist/apps/perion-steam-challenge/.next/types/**/*.ts",
     "next-env.d.ts",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    "./env.d.ts"
   ],
   "exclude": [
     "node_modules",
