+++ libs/steam-api/README.md
@@ -1,0 +1,7 @@
+# steam-api
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test steam-api` to execute the unit tests via [Jest](https://jestjs.io).
