+++ libs/steam-api/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "steam-api",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/steam-api/src",
+  "projectType": "library",
+  "targets": {},
+  "tags": []
+}
