+++ libs/steam-api/src/lib/steam-api.spec.ts
@@ -1,0 +1,5 @@
+describe('steamApi', () => {
+  it('should work', () => {
+    // todo
+  });
+});
