+++ libs/utils/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "utils",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/utils/src",
+  "projectType": "library",
+  "targets": {},
+  "tags": []
+}
