+++ libs/utils/src/lib/utils.spec.ts
@@ -1,0 +1,5 @@
+describe('utils', () => {
+  it('should work', () => {
+    // todo
+  });
+});
