+++ package.json
@@ -44,6 +44,7 @@
     "eslint-plugin-react-hooks": "4.6.0",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
+    "jest-environment-node": "^29.4.1",
     "nx": "18.0.5",
     "prettier": "^2.6.2",
     "sass": "1.62.1",
